Implement and test functionality for arbitrary interface properties
The Description and Capabilities columns have been moved out of
the Interface table and into the Property table.  The Property
Table consists of InterfaceID, Key and Value columns.  The Keys are
integers corresponding to the PropertyKey enum of
QServiceInterfaceDescriptor.

All table columns are now NOT NULL columns.
A superfluous executeQuery function has been removed.
Rollbacks and commits now use the ServiceDatabase's
databaseRollback() and databaseCommit() functions.