Add UserChoice network configuration type.
UserChoice configurations represent a placeholder that is resolved to
an actual configuration when QNetworkSession::open() is called. It
allows the choice of configuration to be deferred until a point when
an asynchronous operations is performed. This allows the platform to
display a dialog to the user without blocking the program.