Testing DatabaseManager +  fixes + permission/invalid db error codes
There were a corner cases concerning defaut services
that weren't properly handled like reseting an external iface(1)
entry in the Defaults table of the user scope database to a valid
iface(2) (if available) when the service implementing that original
iface(i.e. 1) is removed from the system scope database.

Also generate a DBError::NoWritePermissions error code when a write
operation could not be performed.  And produce a
DBError::InvalidDatabase error code when performing an operation
on a corrupt or invalid database file.