Do not bring down interface on QNetworkSession::close() on Windows.
Windows does not have system session support, be conservative and do not
bring down the interface when the last in process session closes, just
in case another process is using the interface.

This change in behaviour allows the session reference counting code to
be removed greatly simplifying the implementation.

QNetworkSession::stop() still forces the network interface down.

The tst_QNetworkSession::sessionOpenCloseStop() unit test should now be
appropriate for testing both in process and system session support.