Tests + fixes for default service edge cases regarding DatabaseManager
When both user and system databases are opened an "integrity check"
can be performed for external interfaceIDs in the user database.
External InterfaceIDs in the user scope db that no longer refer
to interface implementations the system scope db needs to either
be reset to an appropriate default if one can be found or
removed if not.