Refactor warnings on DatabaseManager::open()
Do not produce warnings if the system database does not exist and cannot
be created while a user scope operation is carried out.  It usually
just means the system db has not been created yet and the user
doesn't have write permissiont to create one.

Also only produce one error message on first try if there is a problem
opening one of the databases.  Further function calls
will try to open the database again but the warning is only
output on the first call.