Change: Simplified the code nco:PersonContact query code.
This commit simplifies how we access the LiveNode that is the result
of the query we did earlier to see if the resource exists or not.

We also now implicitly access the LiveNodes as if it would be a QAbstractItemModel
when asking for how many rows it had.

Checking explicitely that the contact exists.

Some small changes to contact().

contact data copying from tracker.

TODO more attributes for contact details.

Avatar reading/writing between qcontact and tracker.

also identified two unittests which causes tracker to need restart.
added QVERFIY(false) at the beginning of these until they are fixed.