Basic implementation of QMessageStore::queryMessages that supports slices (inefficiently), but only supports the default sort and filter keys.
A list of valid QMessageIds is returned, this should be sufficient
to allow work to begin on the QMessage and QMessageContentContainer
implementations.

Also work can begin on QMessageServiceAction, and filter and sort
keys for messages, but that work may conflict more easily with
the work to be done in QMessageStore.
The search is performed by individually searching each folder in
the hierarchy of folders, and copying the entry id and record key
of each message which are use to construct a QMessageId.

There is still lots of debug left in at the moment.

There's documentation in the form of TODOs for work planned.