Add support for mapi entry id look up from mapi record key based message identifiers. Entry ids go stale 'after a few months' so a persistent alternative is required.
Record ids are persistent but not atomic and can not be used as
indexes.

So due to lack of atomicity, for a message identifier it is
necessary to include a mapi store id (store record key is used) and the
record key for the mapi message.

But with only the mapi store and mapi message record keys available
finding the message entry id would require searching all folders in
a store for a message with a matching id. Hence to accelerate
message location by record key based identification the mapi
folder record key is also stored as part of the record key
based identifier.

It's still necessary to iterate over folders to find a folder with
a matching record key, and then search that folder using a mapi
table with an IRestrict set.

Later when multiple stores/accounts are supported, it will also
be necssary to iterate over the account ids, but that should be
simple.