Work around a problem with MWC.
Strange issue with template specializations - unless it is fully
defined MWC chooses to ignore it, but if it is fully defined (in
the header) all other compilers generate duplicate symbols and
cause link problems.  So, for MWC put the definition in the header
and for everything else put it in the cpp file with a declaration
in the header.