Slide show first draft.
Loads images from a playlist at regular intervals.
Should display the cover art for music, or poster for videos if the
playlist can supply a URI.
All I/O through QNetworkAccessManager so it's not limited to the local
file system.
Playlist API is a bit at odds with QMediaPlayer.
No outputs.  Images currently displayed by listening to a signal emitted
by the service.
Doesn't adequately respond to changes in playlist contents.
Timeout starts from when image is requested not after it's loaded so the
display time may be a bit short.