Move meta-data accessors to QAbstractMediaObject and remove QMediaMetadata.
Fleshes out the QAbstractMediaObject class a bit, and removes a
somewhat awkward class (name, and function) in QMediaMetadata.

Adds a registerService() function to QAbstractMediaObject so it can
access the service control.
Makes meta-data names mixed case (metaData, MetaData), which is
consistent with the other metas (MetaObject, MetaProperty, etc).

Rename metadataAvailable() to isMetaDataAvailable() and
metadataAvailabiltyChanged() to metaDataAvailableChanged().

Rename isReadOnly() to isWritable() and invert return value to
better emphasise it's about whether the meta-data can be written.