Fixed async request result handling
* Fixed: Invalid contact ids were handled as valid.

* Testing using just our async request without the manager => possible to
step into the code from unit test.

Fixed the trackerplugin savePhoneNumber test.

The problem was that the affiliation was never saved to tracker, but the
following fix took care of that:

-    RDFFilter doesntExist = affiliation.not_().isBound();// do not create if it already exist
+    RDFFilter doesntExist = affiliation.isBound().not_();// do not create if it already exist

Thanks Johan for pointing this out.

The rest of the commit went into making the code readable.

Fixed ut_qtcontacts_trackerplugin::testSaveName.

Middle name was not read from tracker.