Add: QTrackerContactsLive class
This class abstracts the use of Live nodes when reading and writing
QContacts to Tracker.

NOTE: Only name and nickname are yet saved with this approach!
This approach can be implemented and improved in parallel with the
current implementation to avoid regression.

This is still work in progress, and more of a proposal.