(More) correctly set QMessage::Incoming flag. PR_MESSAGE_HEADERS which was used to detect incoming messages is also set for (all?) my sent mail.
So use PR_RECEIVED_BY_ENTRYID which seems to be only set for incoming
mails (by inspection) to indicate that a message is incoming.

And also use PR_END_DATE to indicate that a message is incoming. This
property is set for meeting invitations.

Ideally and hopefully standardFolders will make the incoming status field
redundant so that it can be removed.