Fixed link of tests/examples on mac in debug mode with frameworks.
On most platforms, debug-and-release libs are put into build/Debug
and build/Release respectively.

This is at odds with how debug-and-release frameworks are supposed
to work: the debug and release libraries go into the same bundle.
Apps are never explicitly linked to the debug or release library,
instead they are linked to the framework, and the debug lib is
selected by setting the DYLD_IMAGE_SUFFIX environment variable at
runtime.

For this to work as expected for mobility projects, we'll make sure
frameworks always go under build/Release.  Previously, the release
half of the framework would go to build/Release and the debug half
would go to build/Debug.