Added expected failures to Location tests on WinCE.
The test farm does not currently have any WinCE images which include a
running copy of FakeGPS, and the Location unit tests rely on some kind
of GPS data being available.

Until new images are created or some method of remotely installing and
running FakeGPS is devised the tests are expected to fail on WinCE,
and have now been marked as such.

This is detailed in issue MOBILITY-337.