Improvements to connectNotify() related tests.
The TestQGeoSatelliteInfoSource::removeSlotForRequestUpdate() test
was requesting an update within 60 seconds and then waiting 60 seconds
for the timeout to occur.  This fails in the presence of data, and was
fixed by requesting an update within -1 seconds, which is guaranteed to
emit the desired timeout signal.

Tests similar to the slot disconnecting tests for
QGeoSatelliteInfoSource were also added for QGeoPositionInfoSource.