Wrap QMessageStore with QMessageManager.
In order to provide a consistent interface with other components
of QtMobility that provide a central data manager interface, the
existing QMessageStore interface is now private, and replaced in
the public interface by the QMessageManager class.

The new class provides exactly the same interface as the old class
excpet that it is instantiated as a handle rather than accessed as
a singleton.  Internally the handle object can be instantiated as
needed, but in documentation or examples the class should be treated
as if it were potentially expensive to create and destroy; this will
give a consistent form of use for all Q{x}Manager classes between
the various QtMobility projects.

The old QMessageStore class is still used internally by the
QMessageManager, so the latter class need not be reimplemented for
multiple platforms.