Rename QMessageServiceAction to QMessageService.
The term 'action' is already used with conflicting meanings;
it can represnt a user-input (QAction), or a facility provided
by a service provider which is exposed to the user (as in
QContactAction).

QMessageService provides a proxy, dispatcher or broker
facility, but adding any of these terms to the class name has
been judged as not constituting an improvement to readability.