Added compile tests for SNAP feature.
If s60::cmmanager headers are present in the build,
configure.bat shall set 'snap_enabled=yes' in the
resulting config.pri - file. Upon detecting this,
the 'bearer.pro' shall define the 'SNAP_FUNCTIONALITY_AVAILABLE=1'
(or not, if that's the case).

In practice SNAP is supported from S60 3rd Edition
Feature Pack 2 onwards.