Fix empty bearer name on some platforms.
An empty string returned by QNetworkSession::bearerName() could mean
either the bearer type is unknown or that there is no bearer (i.e. the
session is invalid).

Distinguish between these two cases.  "Unknown" is returned when there
is a bearer but the bearer type is unknown.  An empty string is only
returned when there is no bearer.

Task-number: MOBILITY-391