Explicitly share QContactDetail, detach when required.
Now that the key in a detail is globally unique, we can share on the
stack without qualms.  However, we don't want changes made by the
client to a detail to propagate automatically to the (previous version
of the) detail which was saved in a contact.  ie, they must call
saveDetail(..) first.  As such, we must explicitly detach on save and
on retrieve of the details.  Finally, we also modify the removal code
so that details are removed based on their key rather than contents.