Suppressed blocking system dialog when calling GetDiskFreeSpaceEx()
The blocking system dialog occurs when checking the free space of an
empty floppy drive with GetDiskFreeSpaceEx(), and requires user
intervention to close. While the blocking system dialog is shown, a
call to GetFileAttributes(), for example by
QFSFileEnginePrivate::doStat() in qfsfileengine_win.cpp, hangs until
the blocking system dialog is closed.

So we should suppress the blocking system dialog to prevent other
functions from hanging.