Undo certain changes made as a result of API review
One of the proposed changes was to replace detailDefinitions() with
detailDefinitionMap() (likewise for other similar signatures).
Review suggests that such a change would have negative impact on the
API, and so this has been undone.

One change was that variantValueMap() deprecated values(), whereas
now values() is still deprecated, but by a variantValues() function.

Finally, the QContactAction::Status enum was deprecated in favour of
QContactAction::State, for consistency with async requests, etc.