Fix unit tests, enable deprecation of QContactName fields
Replace the field key values for QContactName::FieldFirst,
FieldMiddle and FieldLast with the same values that are used
for the (new, replacement) fields FieldFirstName, FieldMiddleName
and FieldLastName.  This enables simple removal of those fields
after the deprecation period has elapsed.

Also updated some unit tests so that all unit tests are now passing.