Clarify requirements on manager (and backends) on contact save.
Previously, it was not clear whether or not the backend had to update
the contact argument to the saveContact() function with the latest
details / information available for that contact.
This change makes it clear that the only two pieces of information
that must be updated by the backend are the id and display label
of the saved contact.  If the client wishes the other fields to be
updated also, they should fetch that contact again by calling
the QContactManager::contact() accessor.