Deprecate progress() signal, modify stateChanged() signal in QCAR
The progress() signal was renamed to resultsAvailable(), since the
waitForProgress() slot was deprecated.  In some cases (eg, contact
remove request) this signal makes no sense (no results) and so
was removed.

Also, a "self" pointer argument was added to the stateChanged()
signal in QContactAbstractRequest so that clients don't have to use
::sender().