Add Capabilities field to QContactOnlineAccount, change request signal
In order to filter contacts by which actions are available for them,
the capabilities which are supported by each online account detail
must be stored in the contact.  The capabilities field is a stringlist
which contains these capabilities.

The various async request leaf classes emit a signal which used to
contain a pointer to itself.  This can result in segfaults (e.g., when
the signal is emitted from a thread which deletes itself, which
is connected via a queued connection to a slot in a different thread).
In that case, using sender() will tell the receiver that the sender
is now null.

Also, some requests had an appendOnly flag (result order unchanged),
which needed removing: can result in receiving an earlier signal
which has the append only flag set, when the real object has newer
results which are reordered.  This flag was removed, and perhaps
in the future the result ordering will become part of the request
state.