Refactor the backend registration.
There is now an interface (QSensorBackendFactory) instead of a function pointer
used to create new backend instances. The plugins have been updated to use this.
The docs have been updated to mention this.

If you do not ask for a specific backend and the default backend is not suitable,
any other registered backends will be tested.