Update the symbian async implementation, remove usage of depr. API
The unstable branch contains new API which deprecates significant
portions of the old manager API.  This change removes some of these
usages of deprecated API in the symbian and wince engines.

The new async implementation in the symbian engine is now the same as
that in the memory engine, with the definition save and remove
request implementations removed (as it doesn't support mutable defs).