Separate state and result update functions in QContactManagerEngine
Also move the resultsAvailable() signal into the base request class
now that all leaf classes emit it.  Update the documentation for the
leaf request classes to reflect the situations in which the
resultsAvailable() signal will be emitted.

The separation of state update and result update functions will enable
engines to call the appropriate engine trampoline function only when
required.