Remove setDetailOrder fn, deprecate preference fns in QContact
The setDetailOrder function is not supportable, and should not have
been merged in during the recent unstable merge.

The preference functions (setting a particular detail as the preferred
detail for a particular action) are also difficult to support
(as it would require backends to listen for changes to the available
actions, and save per-contact, per-detail preferences in-band).
These were deprecated, and will be removed after the transition period
has elapsed.