Changes: Add unit test to show that fetching works in general.
RevBy: people-review
Details:
* tests/ut_qtcontacts_fetch/ut_qtcontacts_fetch.[h|cpp|pro]:
Add a test that saves, checks (via fetch), and removes a contact.
It checks that _any_ contact is fetched, not that the specific contact
is fetched, though that can be changed by changing one bool in doNextOperation.

Changing that bool makes this do the same as ut_qtcontacts_add_async(),
but this test case also experiments with a state enum instead of the
nasty callback functions. I have not decided which is best for future unit tests.