d-pointerize the map object classes
Also rename QMapObject::QMapObjectType enums to have Object suffix to
follow convention that enums should have at least one element of enum
type name in each of the enum values.

QMapObject's private implementation now has a const QMapView member
pointer, rather than a reference. The proposed future changes to
the map objects are that they can be constructed and then assigned
to a QMapView.  A member reference would not work in this case
since it has to be initialized on construction.

Also provide implementations for QMapEllipse and QMapPixmap.

Adjusted documentation from doxygen to qdoc, there still remains
some documentation work to do.

Some functions such as QMapObject::type(), ::pen() functions
etc are no longer inline.  Inlining these classes would necessitate
including the private headers in the public headers which is
generally not advisable.  There may be reasons to reverse
this in future if inlining is necessary.