Rename qaddress to qgeoaddress, d-pointerize and add set functions
The rename was to make it more obvious that the address was map/geo
related.

Set functions were added, when people create landmarks they will need
to be able to set the address.  Consequently QGeocodingXmlParser
no longer needs to be a friend class and has been appropriately
updated.

Also added in a post office box field to the qgeoaddress.