d-pointerize qgeolocation and some pruning.
Remove title and language fields which seem
superfluous.  The "title" according to the rest documentation
is effectively a string containing the address.  The language
field describes language of this "title".

A more appropriate place for title and language is qgeoaddress,
but serves no purpose because qgeoaddress already has
the address data albeit split up.

Also remove the alternative address too.  If someone wants an address
in a certain language, then they can set the language in the request.
It does not make sense to return all the alternative translations
of an address everytime we make a reverse geocoding request.

QLandmark has been modified to use the QGeoLocation class
as a member.  Also use a QSharedDataPointer for the QLandmark
d pointer.