Implemented support for structured property value types in QtVersit
Some properties in vCards have internal structure, eg. N or CATEGORIES.
Previously, the structure was ignored and simply passed into QVersitProperty
as an opaque string.  This led to problems in parsing.

This commit makes the reader, writer, importer and exporter aware of this
structure and fixes bugs in parsing using the new framework.

Relates to QTMOBILITY-89