Fix for QTMOBILITY-65.
This adds documentation of the Symbian capabilities required by
each of the Mobility modules.  Note that the capabilities listed may
in fact be overkill for common use cases.  Hopefully the module
maintainers will add or improve Symbian capability documentation to the
main page for their modules in time.

Also added the regenerated docs.