Refactor paint algorithms of map objects.
Removed duplication of code involved when objects cross the dateline.
We do this by letting the QMapViewPrivate::paintLayers(),
handle the situation instead.

Make use of the painter's save and restore operations so saving
and resetting of brushes/pens isn't needed.

Change the behaviour of QMapLine to draw always draw the shortest
path between two points.  And also move the connectShortest() function
out of qmapview and into qmapline.