Don't check the timestamp when updating contact in maemo5 backend
Since the contact is a value class, we shouldn't check the timestamp
of the contact during a save operation to ensure that it is a modified
version of the most up-to-date contact in the database.

Instead, we should just overwrite the existing contact in the database
as per the documented behaviour.

Relates to task QTMOBILITY-170