Branch filtering test on supported schema of contacts backend
Some backends don't support certain fields or detail types which
are used by the filtering tests.  On such backends, don't perform
the tests which use those unsupported fields/details.

Also, the maemo5 backend strips certain characters from phone numbers
on save.  This change ensures that we don't do exact match filtering
on that might change, by changing the phone number that was saved in
the test contacts so that they do not contain a hyphen.

Relates to task QTMOBILITY-92