Minor fixes for QLandmarkManager
Add in constructor that only accepts a parent as a parameter.  This
allows the user to create default instance and provide a parent
if desired.

Add in overloads for landmarks() and landmarkIds() that accepts
a single sort order.  The primary use case is for a single sort
order and so makes it more convenient for the user.

Provide a default parameter for categories().  It allows a user
to conveniently request call categories.