Fix for failing requests in S60 Location backend.
Failing unit tests indicated that no requests (position or satellite)
were being made on S60.

Further investigation showed that the fix time metrics had changed
such that none of the position modules were acceptable as per the
current checks on the modules.

This was fixed by adding a second round of checks for this case,
which searches for the module with the best fix time metric rather
than just for fix time metrics less than the timeout.

It also became apparent that a small change made concurrently with
the last round of changes addressing unit test pass rates added a bug
which caused the device to panic on negative timeouts and to behave
incorrectly for timeouts less that the minimum acceptable timeout.

This was fixed as well.