Flesh out landmark request classes.
In deviation from contacts, we force a manager to be passed into
the constructors.  A manager is always required and we can't have
a default manager being built everytime a request is created.
There's also issues with deletion  etc.

Also filters and sort orders are passed in as pointers rather
than references.