Change the QGalleryUrlRequest API to only return an itemId and itemType.
Unlike other requests the type and supported properties of a URL cannot
be determined in advance and therefore would require dynamic changes
to the available properties, and property attributes.

Secondly it would appear that in most use cases for the URL request a
two request query would be used anyway, one to determine the type and
a second to request a set of properties appropriate for viewing items
of that type.

Finally this makes the simple use case clearer, and reinforces the URL
request as an entry point into the gallery as a separate enitity rather
an extension of the file-system.