Add isReadOnly functions into manager and engines
Add functions to see if the entire manager or whether certain
categories and landmarks are read only.

remove isReadOnly() from categories, it make little sense to
have a member function do this since we might want to
have a category that is read only in one manager saved to a different
manager.