Improved interface QVersitContactImporterPropertyHandler
Changes to QVCIPropertyHandler::postProcessProperty():
Added parameter QList<QVersitProperty>* updatedDetails as a way for the
handler to tell what properties the base exporter generated and to
modify them

This was added as a new class (QVCIPropertyHandler) for source/binary
compatibility.

Deprecated detailHandler() getter as this is now ambiguous with multiple
handler versions.