Remove implementation of setItemDisplayLabel, modify schema.
The schema for each of the various item types has been expanded to a
minimally useful state.  The snippets example now runs with the memory
backend and shows item polymorphism.

The implementation of setItemDisplayLabel did set the zeroth detail,
which is currently the item type, hence this was a bug.  Currently,
there is no implementation for setItemDisplayLabel; this is TODO.