Symbian: Performance improvement to contact fetches.
If there is a QContactFetchHint with a detail definitions hint parameter
defined when calling QContactManager::contacts symbian backend now uses a
view definition parameter when calling CContactDatabase::ReadContactL.
This means that only the defined contact fields are read from the database.
Also, the check for fetch hint during the contact transform phase was
removed as it is now obsolete. These changes decrease the time required for
contact fetches about 30% (tested with 500 and 1000 contacts).