Add more code to allow generation of instances in memory backend
Uses the previously written functions to generate instances within a
time period for each xrule and rrule of the generating item, and
returns them in a list.

Checking that the resultant items are within the correct period is not
currently done properly, and should be revisited once we have some
unit tests for this functionality.