Clarify dataRate.
Also outputRange (which has similar portability issues).

Change the meaning of dataRate/outputRange "defaults" and
stop the backends from setting defaults so that the changes
can have some kind of effect.

Document that the default dataRate(0) and outputRange(-1) mean
the app does not care (ie. in the default case the app will not
attempt to set the dataRate or outputRange).

Remove checks that are no longer valid because of this.
Timer-based loops (N900, dummy, maemo6) are set to the first value
of the first entry in the availableDataRates list.

Fixes: MOBILITY-863