Changed the plugin interface for geoservices.
The plugin now acts more like a factory, which should avoid a known
bug with the previous method.

Additionally, the various managers are now lazily loaded by
QGeoServiceProvider and the plugins parameters are passed into the
managers via a constructor - at the moment it's the plugin's subclass
constructor, the base classes should probably get a similar constructor.