Symbian: Fixed signal emissions for pre 10.1 platforms.
Added a new flag to variate signal emission behaviour between pre 10.1 platforms
and newer. Getting the signal filter (m_contactsEmitted at CntSymbianDatabase)
to work correctlty in all situations proved near impossible. So removing it was
the only option to get this working.

For perfomance reasons 10.1 platform still uses the signal filter but will not
work correctly in all situations. It is under investigation if the signal filter
could be removed also from there.