Use QWeakPointer instead of QPointer in manager engine code
Lighter weight than QPointer, and we only use it to check for deletion
of the QObject in the single threaded case.  We depend on the engine
implementation to provide correct locking to prevent deletion of
requests from occurring until the function has completed, if the
engine has a multithreaded implementation.

IE, in the multithreaded implementation case, the request may not
be deleted between the two emit statements in each of the
updateFooRequest functions, because the engine should lock a mutex
to prevent the request from being deleted until the updateFooRequest
function has completed (ie, locking in requestDeleted()).