MOBILITY-965
On Symbian^3, opening a QNetworkSession based on SNAP that is emtpy
triggers a access point selection dialog. This task modified QtBearer to know
how to react on that (in particular if a new access point was created from
the dialog, or an existing access point was chosen from another SNAP).