Draft of proposal for multiple calendar support
Introduce the idea of an "item collection" which has an id and some
metadata.  Items then have a detail which contains the ids of each
collection to which the item belongs.

Collections are therefore top-level items, saved and removed directly
from managers, whereas the item adding/removing is done via detail
manipulation and item saving.