Add test for removal of details from persisted contacts on update
Previously, only addition of details to persisted contacts on update
was tested, leaving an important piece of functionality entirely
untested.  This commit extends the ::update() test to ensure that
if a detail is removed from a contact, and that contact is saved,
the persisted version of the contact also has that detail removed.