Merge the QGalleryContainerRequest into QGalleryFilterRequest.
The only differences between these two requests were that the container
request did not have a filter argument which it perhaps should of, and
it returned only the direct descendents of its container item, which
can be solved with an extra argument to the filter request.

The count request also gets this new argument which means it can count
both cases.