Doc changes and some routing changes ahead of API discussion.
Mostly adding documentation.

I commented out some bits and pieces that were added to route request
which probably belong elsewhere.  I'm preparing the API for review so
any changes to the routing / places / non-tiled mapping API between
now and the API review will probably get rolled back, since we need
stable header files and documentation for the reviewers to browse (
and I need to check the doc/API we have in the meantime).

There are comments in the code describing what was commented out, but
I'll mention them here.

I don't see how the DirectRoute route optimization is any different
from the ShortestRoute optimization.  I know there's a DirectDrive
route optimization with the Nokia Services, but that's something
that would only be exposed in a Nokia specific API.

The time of arrival / time of departure information probably belongs
elsewhere - possibly in a transit options subclass once those come into being.
I'm hoping we'll be able to get all of the time/traffic aware options
into a transit options class and have most of the rest of the time/traffic
behaviour dealt with either internally to the route objects or at worst
with routesegment / navigationinstruction subclasses.