Fix bug in maemo5 backend affecting contact save operations
The backend really has uniqueness constraints on every type of detail
other than QContactPhoneNumber and QContactOnlineAccount.

This patch ensures that these constraints are reported to the user,
and that the appropriate error is reported when attempting to save
multiple unique details in a single contact.

It also adds logic to make non-uniqueness work correctly for details
of QContactPhoneNumber type.

Relates to task QTMOBILITY-310