Updated the plugin interface for Maps.
This involved a Plugin->Factory name change, exposing and using
verison numbers, and some general API cleanups to make it harder to
misuse.

For example, the QGeoServiceProvider now sets the
manager name and version from the provider name and version, and the
managers are constructed from the engines in the QGeoServiceProvider
classes as well.

This is done through private setters / constructors to avoid confusing
the API and associated doc for clients / implementers.

The managers and engines are now friends of QGeoServiceProvider, which
is less nice than I'd like, but it seems to be a win overall.