Fix uniqueness constraints in maemo5 backend
We now have email addresses and street addresses, as well as phone
numbers and online accounts (and presence, once implemented) as
non-unique definitions in the maemo5 engine schema.

This patch also ensures that all details are removed from a contact
and then re-added on save, so that no delta/modification set has
to be calculated and applied, simplifying the save process.

Relates to tasks QTMOBILITY-310 and QTMOBILITY-325