Implemented logic in memory engine to fix Occurrences on save
When the importer imports items, it cannot set the parentLocalId() on Occurrences properly to refer
to the parent event.  This responsibility is therefore that of the engine - to check the Guid and
the ParentLocalId on Occurrence types on item save and ensure they are consistent.  This is such an
implementation for the memory engine.