Allow action implementers to implement supportsContact()
Previously, it was hardcoded to return the result calculated naively
from the fetched list of supported targets for a contact.

An action implementer may be able to do this in some faster way,
so that function is now virtual.

The action descriptor implementation of this function has been changed
to call the factory's implementation of the function, also.