Remove action descriptor accessor from action factory
The action descriptor can be fully specified by the tuple of the
action name, the service name, the implementation version and the
pointer to the factory which instantiates the actions.

Thus, there is no need for the factory to have to provide this, as
these pieces of information are all available to the
QContactActionServiceManager.