Fix the action filtering tests
Previously, the test used a static map of manager to detail definition
and field to determine which def/field to use for a particular test
which would be suitable for a particular action.

This no longer works, so we have to hardcode the definition name and
field key for each action.  Thus, the only manager which this is
suitable for is the memory manager (since we have to add arbitrary
definitions which support the appropriate data type).

So, this patch adds in special case code to determine whether the
manager under test is the memory manager, and if so, prepare the model
as required and execute the tests.