remove QAbstractAnimation as the base class
The interfaces are thus a little bit simplified.
The error reporting is now entirely dependent on the backend to call
the reportError function. It allows for async errors as well.
The state function is still there but it is pull only. There is
no signal notifying a change in the state because it just takes
too much CPU time and isn't accurate at all.