Begin Refactoring and test importing of landmarks
The code responsible for importing has been separated out from
QLandmarkManger because we need to be able to handle import operations
asynchronously. QLandmarkManager is a qobject and so only has affinity
to one thread.

For the same reasons, the file handler classes no longer accept a
manager engine as a parameter so that we can implment the aysnchronous
case with multiple threads.

The category handling code for the lmx file handler and tests
have been commented out.  The lmx schema only includes categories
for a landmark if the categories are global categories.  This
is something which need to be implmented properly further down the line.

Need to refactor landmark saving to use placeholder binding.  Previously
coordinate values were being truncated because they were converted
to strings and appended into the sql query.  Place holder binding
also prevents sql injection.